requirements.txt
alembic==1.2.1
astroid==2.3.2
autopep8==1.4.4
Babel==2.7.0
bcrypt==3.1.7
blinker==1.4
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cryptography==2.8
dj-database-url==0.5.0
Django==2.2.6
django-cors-middleware==1.4.0
django-heroku==0.3.1
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
djangorestframework-oauth==1.1.0
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.5.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
logilab-common==1.4.4
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
paramiko==2.6.0
passlib==1.7.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.3
PyNaCl==1.3.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
simplejson==3.16.0
six==1.12.0
speaklater==1.3
SQLAlchemy==1.3.10
sqlparse==0.3.0
sshtunnel==0.1.5
typed-ast==1.4.0
urllib3==1.25.6
Werkzeug==0.16.0
whitenoise==4.1.4
wrapt==1.11.2
WTForms==2.2.1
